% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_summary.R
\name{glm_summary}
\alias{glm_summary}
\title{A concise (if incomplete) summary of glm results}
\usage{
glm_summary(model)
}
\arguments{
\item{model}{This is a glm model.}
}
\value{
data.frame of Odds Ratios, 95% confidence intervals, and p-values
}
\description{
Rather than describe the model adequately, "cut to the chase"
}

