% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nsqip_readmitted.R
\name{nsqip_readmit}
\alias{nsqip_readmit}
\title{ACS-NSQIP Readmission identification}
\usage{
nsqip_readmit(df)
}
\description{
There are several variables to identify if a patient
 was readmitted or returned to the hospital. This function
 identifies if the patient was readmitted.
}
\details{
@param df This is the NSQIP dataframe

@export

@return Characters, "Readmitted" if readmitted and "Not Readmitted" if not.
}

