% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wide_screen.R
\name{wide_screen}
\alias{wide_screen}
\title{Wide screen}
\usage{
wide_screen(howWide = Sys.getenv("COLUMNS"))
}
\arguments{
\item{howWide}{Accepts numeric input, but it's better to
 let the function do its work

 @example wide_screen()

 @export}
}
\description{
Adjust R output to current terminal width (UNIX)
}

